+++ HISTORY.md
@@ -11,6 +11,7 @@
 ### 2021-05-11
 * gui: add multi-scene projects 
 * gui: adjusted scene album rendering
+* gui: tweaked scene album rendering
 
 ### 2021-05-10
 * gui: project menu entries for "new scene" and "add scene"